user authentication updated

+++ app/Repositories/TodoRepository.php
@@ -3,21 +3,30 @@ namespace App\Repositories;
 
 use App\Interfaces\TodoInterface;
 use App\Models\Todo;
+use Illuminate\Foundation\Mix;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
+use Illuminate\Pagination\LengthAwarePaginator;
+
 
 class TodoRepository implements TodoInterface{
 
-    public function fetchAll($request): Collection
+    public function fetchAll(int $perPage, string $sort_by, string $order, string $filter): LengthAwarePaginator
     {
-        return Todo::where(function ($query) use ($request) {
-            $query->where('status', $request->status);
-        })->get();
+        return Todo::when($filter != null ,function ($query) use ($filter) {
+            $query->where('status', $filter);
+        })
+        ->orderBy($sort_by, $order)
+        ->paginate(10);
     }
 
-    public function fetch(int $id): Todo
+    public function fetch(int $id): mixed
     {
-        return Todo::findOrFail($id);
+        $Todo = Todo::findOrFail($id);
+        if($Todo->user->id == auth()->user()->id){
+            return $Todo;
+        } 
+        return null;
     }
 
     public function create($data): Todo
@@ -25,18 +34,24 @@ class TodoRepository implements TodoInterface{
         return Todo::create($data);
     }
 
-    public function update($data, int $id): Todo
+    public function update($request, int $id): mixed
     {
         $Todo = Todo::findOrFail($id);
-        // dd($data);
-        $Todo->update($data);
-        return $Todo;
+        if($Todo->user->id == auth()->user()->id){
+            $Todo->update($request->all());
+            return $Todo;
+        } 
+        
+        return null;
     }
 
     public function delete(int $id): bool
     {
         $Todo = Todo::findOrFail($id);
-        $Todo->delete();
-        return true;
+        if($Todo->user->id == auth()->user()->id){
+            $Todo->delete();
+            return true;
+        } 
+        return false;
     }
 }
